requirements.txt
numpy==2.1.1
opencv_contrib_python==4.10.0.84
Pillow==10.4.0
streamlit==1.38.0
